Changes for NixOS 24.05
program.gnupg.agent.pinentryPackage should be fine with the default
value, but I'd rather not assume it will figure out the right value as
I've had this break before.

power.ups.upsmon.settings.RUN_AS_USER is not marked as default in the
module so it needs to be forced. The force can be removed once
https://github.com/NixOS/nixpkgs/issues/318582 is closed.